+++ src/main/java/dev/linkcentral/presentation/BaseUrlUtil.java
@@ -1,0 +1,12 @@
+package dev.linkcentral.presentation;
+
+import javax.servlet.http.HttpServletRequest;
+
+public class BaseUrlUtil {
+    public static String getBaseUrl(HttpServletRequest request) {
+        return String.format("%s://%s:%d",
+                request.getScheme(),
+                request.getServerName(),
+                request.getServerPort());
+    }
+}
