+++ src/main/java/dev/linkcentral/presentation/controller/view/DefaultViewController.java
@@ -1,4 +1,4 @@
-package dev.linkcentral.presentation;
+package dev.linkcentral.presentation.controller.view;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -8,10 +8,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 @Slf4j
 @RequiredArgsConstructor
-public class IndexController {
+public class DefaultViewController {
 
     @GetMapping("/")
-    public String Home() {
+    public String home() {
         return "/home";
     }
 }
